refactor: 등급별 적립률 gradePointRatio 타입 변경 (Integer -> BigDecimal)
정밀한 계산을 위해 Grade 엔티티의 gradePointRatio 타입을 Integer에서 BigDecimal로 변경.

+++ src/main/java/com/nhnacademy/memberapi/dto/request/GradeRequest.java
@@ -2,8 +2,10 @@ package com.nhnacademy.memberapi.dto.request;
 
 import com.nhnacademy.memberapi.entity.GradeName;
 
+import java.math.BigDecimal;
+
 public record GradeRequest(
         GradeName gradeName,
-        Integer gradePointRatio,
+        BigDecimal gradePointRatio,
         Integer gradeCondition
 ){}
